GridCellEncoder: Move hexy to nupic.external
