+++ py/src/nupic/external/hexy.py
